+++ src/core/contexts/taskStoreContext.tsx
@@ -1,9 +1,6 @@
-import React, { createContext, useContext, useEffect, useRef } from 'react';
-import { useLocation } from 'react-router-dom';
+import React, { createContext, useContext, useRef } from 'react';
 
-import { useQueryClient } from '@tanstack/react-query';
 import { create } from 'zustand';
-import type { QueryClient } from '@tanstack/react-query';
 
 interface TaskState {
   overriddenTaskId?: string;
@@ -37,7 +34,6 @@ export const createTaskStore = () =>
 const StoreContext = createContext<ReturnType<typeof createTaskStore> | null>(null);
 
 export function TaskStoreProvider({ children }: React.PropsWithChildren) {
-  useInstantiationUrlReset();
   const storeRef = useRef<ReturnType<typeof createTaskStore>>(undefined);
   if (!storeRef.current) {
     storeRef.current = createTaskStore();
@@ -52,30 +48,3 @@ export const useTaskStore = <T,>(selector: (state: TaskState) => T) => {
   }
   return store(selector);
 };
-
-/**
- * When the URL changes (possibly because we redirected to the instance or because the user went
- * back to party selection after getting an error), we should clear the instantiation. We do this so that we're ready
- * for a possible next instantiation (when the user comes back to try instantiation again). This is needed
- * because the components that trigger instantiation might do so repeatedly - and we need to stop that.
- *
- * Some places instantiate as a direct result of a user action (clicking a button). Those will force
- * re-instantiation anyway and won't care if the previous instantiation was cleared before trying again.
- */
-export function useInstantiationUrlReset() {
-  const location = useLocation();
-  const queryClient = useQueryClient();
-
-  useEffect(() => {
-    // Reset instantiation mutations when navigating away from instance pages
-    // This prevents duplicate instantiation attempts
-    if (!location.pathname.includes('/instance/')) {
-      removeMutations(queryClient);
-    }
-  }, [location.pathname, queryClient]);
-}
-
-function removeMutations(queryClient: QueryClient) {
-  const mutations = queryClient.getMutationCache().findAll({ mutationKey: ['instantiate'] });
-  mutations.forEach((mutation) => queryClient.getMutationCache().remove(mutation));
-}
